add: basic exception handler

+++ src/main/java/br/com/nobre/commons/dto/Page.java
@@ -7,10 +7,10 @@ public class Page<T> {
 	public final List<T> data;
 	public final int start;
 	public final int limit;
-	public final int size;
-	public final int totalPage;
+	public final long size;
+	public final long totalPage;
 	
-	public Page(List<T> data, int start, int limit, int size, int totalPage) {
+	public Page(List<T> data, int start, int limit, long size, long totalPage) {
 		this.data = data;
 		this.start = start;
 		this.limit = limit;
